Refactor code by removing local Verifier classes and importing them from settingsVerification module
